# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Test Go ClickHouse Middle",
        "title": "Test Go ClickHouse Middle",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/event": {
            "post": {
                "description": "insert event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Insert Event",
                "operationId": "create-event",
                "parameters": [
                    {
                        "description": "event info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.successResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Event": {
            "type": "object",
            "properties": {
                "EventTime": {
                    "type": "string",
                    "example": "2023-04-09 13:00:00"
                },
                "Payload": {
                    "type": "string",
                    "example": "{\"some_field\":\"some_value\"}"
                },
                "UserID": {
                    "type": "integer",
                    "example": 1
                },
                "eventID": {
                    "type": "integer",
                    "example": 1
                },
                "eventType": {
                    "type": "string",
                    "example": "login"
                }
            }
        },
        "http.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "http.successResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}